config Path, Controller Auth

+++ src/main/java/com/example/mndc/controller/BoardController.java
@@ -1,19 +1,26 @@
 package com.example.mndc.controller;
 
+import com.example.mndc.auth.PrincipalDetails;
 import com.example.mndc.model.dto.BoardDTO;
 import com.example.mndc.model.dto.MilitaryUnitDTO;
 import com.example.mndc.service.BoardService;
+import com.example.mndc.service.UserService;
+import com.example.mndc.util.BoardPath;
+import com.example.mndc.util.UserPath;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
 @RequestMapping("/board")
-public class BoardController {
+public class BoardController implements BoardPath, UserPath {
 
     @Autowired
     BoardService boardService;
+    @Autowired
+    UserService userService;
 
     @GetMapping("/lists")
     public String viewBoard(@RequestParam(defaultValue = "1") int type,
@@ -21,7 +28,7 @@ public class BoardController {
         // 부대 정보를 보여줄 컨트롤러 입니다.
         model.addAttribute("list",boardService.getUnitInfo(type));
 
-        return "/board/lists";
+        return BOARD+LISTS;
     }
 
     @PostMapping("/lists")
@@ -39,7 +46,7 @@ public class BoardController {
                           @RequestParam(defaultValue = "1") int page, // 페이지 번호
                           Model model){
         model.addAttribute("list",boardService.getBoardsInfo(mu_pk,page));
-        return "/board/list";
+        return BOARD+LIST;
     }
     // 게시글 보기
     @GetMapping("/view")
@@ -49,39 +56,47 @@ public class BoardController {
                           Model model){
         model.addAttribute("list",boardService.getBoardsInfo(mu_pk,page));
         model.addAttribute("board",boardService.getBoardInfo(b_pk));
-        return "/board/view";
+        return BOARD+VIEW;
     }
     // 글쓰기
     @GetMapping("/write")
-    public String writeBoard(){
-        // TODO : 필터를 통해 로그인이 안 됐을시 접근 제어
+    public String writeBoard(@AuthenticationPrincipal PrincipalDetails principalDetails){
+        if(userService.isLogin(principalDetails)) {
+            return REDIRECT+LOGIN;
+        }
 
-        return "/board/write";
+        return BOARD+WRITE;
     }
     @PostMapping("/write")
     public String doWrite(BoardDTO boardDTO){
         boardService.writeBoard(boardDTO);
-        return "redirect:/view";
+        return REDIRECT+VIEW;
     }
     // 수정하기
     @GetMapping("/modify")
     public String modifyBoard(@RequestParam int b_pk,
-                              Model model){
-        // TODO : 필터를 통해 로그인이 안 됐을시 접근 제어. 작성자인지 확인하기
+                              Model model,
+                              @AuthenticationPrincipal PrincipalDetails principalDetails){
+        if(userService.isLogin(principalDetails)) {
+            return REDIRECT+LOGIN;
+        }
         model.addAttribute("board",boardService.getBoardInfo(b_pk));
-        return "/board/modify";
+        return BOARD+MODIFY;
     }
     @PostMapping("/modify")
     public String doModify(BoardDTO boardDTO){
         boardService.modifyBoard(boardDTO);
-        return "redirect:/view";
+        return REDIRECT+VIEW;
     }
 
     @PostMapping("/delete")
-    public String doDelete(BoardDTO boardDTO){
-        // TODO : 필터를 통해 로그인이 안 됐을시 접근 제어. 작성자인지 확인하기
+    public String doDelete(BoardDTO boardDTO,
+                           @AuthenticationPrincipal PrincipalDetails principalDetails){
+        if(userService.isLogin(principalDetails)) {
+            return REDIRECT+LOGIN;
+        }
         boardService.deleteBoard(boardDTO);
-        return "redirect:/list";
+        return REDIRECT+LIST;
     }
     
     // TODO : 추천버튼
